Solution to Input Two Floats and Print Their Product

+++ Input Two Floats and Print Their Product@@6788d5012db25efec946f93e/code.c
@@ -2,7 +2,7 @@
 int main(){
     float a,b,product;
     scanf("%f %f",&a,&b);
-    product=(float)a*b;
+    product=a*b;
     printf("Product: %.2f",&product);
     return 0;
 }
